fix(audit): split AccessAudit.php — one class per file for PSR-4
PSR-4 autoloading expects each class to live in a file matching its name.
The single AccessAudit.php declared the Provenance enum plus AccessGrant,
UserAccessReport, ScopeAccessEntry, ScopeAccessReport, SensitiveRoleCell —
none of which the autoloader could resolve in prod without a generated
classmap. Split each into its own file under Audit/.
