+++ netBS/core/SecureBundle/Audit/ScopeAccessReport.php
@@ -1,0 +1,17 @@
+<?php
+
+declare(strict_types=1);
+
+namespace NetBS\SecureBundle\Audit;
+
+use NetBS\FichierBundle\Mapping\BaseGroupe;
+use NetBS\SecureBundle\Mapping\BaseRole;
+
+final readonly class ScopeAccessReport
+{
+    /** @param ScopeAccessEntry[] $entries */
+    public function __construct(
+        public BaseGroupe|BaseRole $scope,
+        public array $entries,
+    ) {}
+}
